Refactor reactive server logic into new `ReactiveHandler`
